admin academics completed and faculty layout ui created

+++ Server/controllers/academicController.js
@@ -4,6 +4,7 @@ import AcademicSession from "../Models/AcademicSession.js";
 import CourseOffering from "../Models/CourseOffering.js";
 import Faculty from "../Models/Faculty.js";
 import Department from "../Models/Department.js";
+import mongoose from "mongoose";
 
 // ========== Academic Sessions ==========
 export const getAcademicSessions = async (req, res) => {
@@ -60,7 +61,8 @@ export const getCourseOfferings = async (req, res) => {
     const offerings = await CourseOffering.find()
       .populate("course")
       .populate("academicSession")
-      .populate("faculty");
+      // .populate("faculty");
+      
     res.status(200).json(offerings);
   } catch (err) {
     console.error("Error fetching course offerings:", err);
@@ -82,6 +84,11 @@ export const createCourseOffering = async (req, res) => {
 export const deleteCourseOffering = async (req, res) => {
   try {
     const { id } = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({ message: "Invalid course offering ID." });
+    }
+
     await CourseOffering.findByIdAndDelete(id);
     res.status(200).json({ message: "Course offering deleted successfully" });
   } catch (err) {
